Copy mock-seating-plan to ticket/select page
