+++ components/common/button.tsx
@@ -25,28 +25,33 @@ export const FancyButton: React.FC<{
 };
 
 export const RegularButton: React.FC<{
-  variant: "black",
+  variant: "black" | "white",
   buttonClass: string,
   buttonProps?: React.ButtonHTMLAttributes<HTMLButtonElement>,
   children: ReactNode,
   onClick?: () => void,
 }> = ({ variant, buttonClass, children, onClick, buttonProps }) => {
-  // Only black variant for now
   return (
     <div className="w-full h-full flex items-center justify-center p-2">
       <button
         className={`
-          relative bg-[#222222] ring-[#222222]/50 group transition-all duration-150
+          relative group transition-all duration-150
+          ${variant === "black" ? "bg-[#222222] ring-[#222222]/50" : "bg-[#EEEEEE] ring-[#EEEEEE]/50"}
           ${buttonProps?.disabled ? "cursor-progress" : "cursor-pointer active:ring-8"}
           ${buttonClass}
         `}
         onClick={onClick}
         {...buttonProps}
       >
-        <span className="relative z-10 text-foreground">
+        <span className={`relative z-10 ${variant === "black" ? "text-foreground" : "text-background"}`}>
           {children}
         </span>
-        <span className="absolute inset-0 overflow-hidden pointer-events-none bg-transparent group-hover:bg-white/10 duration-300" />
+        <span
+          className={`
+            absolute inset-0 overflow-hidden pointer-events-none bg-transparent
+            ${variant === "black" && "group-hover:bg-white/10"} duration-300
+          `}
+        />
       </button>
     </div>
   );
